# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface VideoUploadInput {
  video: Blob;
}

export interface VideoMetadata {
  durationSeconds: number;
  width: number;
  height: number;
  /** @nullable */
  fps: number | null;
  hasAudio: boolean;
}

export interface VideoUploadResult {
  videoId: string;
  metadata: VideoMetadata;
}

export type SegmentAnalysisLabelsItem = typeof SegmentAnalysisLabelsItem[keyof typeof SegmentAnalysisLabelsItem];


export const SegmentAnalysisLabelsItem = {
  damaged_area: 'damaged_area',
  before_repair: 'before_repair',
  repair_process: 'repair_process',
  tools_action: 'tools_action',
  close_up_detail: 'close_up_detail',
  satisfying_moment: 'satisfying_moment',
  final_result: 'final_result',
  vehicle_reveal: 'vehicle_reveal',
  painting: 'painting',
  sanding: 'sanding',
  polishing: 'polishing',
  dent_repair: 'dent_repair',
  uninteresting: 'uninteresting',
  repetitive: 'repetitive',
  blurry: 'blurry',
} as const;

export interface SegmentAnalysis {
  segment_id: string;
  start: number;
  end: number;
  summary: string;
  visual_events: string[];
  labels: SegmentAnalysisLabelsItem[];
  quality_score: number;
  interest_score: number;
  repair_relevance: number;
  transformation_value: number;
}

export interface VideoAnalysisArtifacts {
  segments_json: string;
  analysis_json: string;
}

export interface VideoAnalysis {
  video_id: string;
  duration_seconds: number;
  width: number;
  height: number;
  has_audio: boolean;
  scene_change_seconds: number[];
  segments: SegmentAnalysis[];
  artifacts: VideoAnalysisArtifacts;
}

export interface ReelEditClip {
  segment_id: string;
  source_start_seconds: number;
  source_end_seconds: number;
  order: number;
  selection_reason: string;
  caption_text: string;
  caption_start_seconds: number;
  caption_end_seconds: number;
  reframe_x: number;
  reframe_y: number;
}

export interface ReelEditPlan {
  story_arc: string;
  hook_text: string;
  hook_duration_seconds: number;
  clips: ReelEditClip[];
}

export interface ProcessedReelOutputArtifacts {
  edit_plan_json: string;
}

export type ProcessedReelOutputMimeType = typeof ProcessedReelOutputMimeType[keyof typeof ProcessedReelOutputMimeType];


export const ProcessedReelOutputMimeType = {
  'video/mp4': 'video/mp4',
} as const;

export interface ProcessedReelOutput {
  reel_id: string;
  mime_type: ProcessedReelOutputMimeType;
  width: number;
  height: number;
  duration_seconds: number;
  file_size_bytes: number;
  download_url: string;
  artifacts: ProcessedReelOutputArtifacts;
}

export interface ProcessedReelResult {
  video_id: string;
  analysis: VideoAnalysis;
  edit_plan: ReelEditPlan;
  output: ProcessedReelOutput;
}