# lib/api-zod/src/generated/types/processedReelOutput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProcessedReelOutputArtifacts } from './processedReelOutputArtifacts';
import type { ProcessedReelOutputMimeType } from './processedReelOutputMimeType';

export interface ProcessedReelOutput {
  reel_id: string;
  mime_type: ProcessedReelOutputMimeType;
  width: number;
  height: number;
  duration_seconds: number;
  file_size_bytes: number;
  download_url: string;
  artifacts: ProcessedReelOutputArtifacts;
}