# lib/api-zod/src/generated/types/processedReelResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProcessedReelOutput } from './processedReelOutput';
import type { ReelEditPlan } from './reelEditPlan';
import type { VideoAnalysis } from './videoAnalysis';

export interface ProcessedReelResult {
  video_id: string;
  analysis: VideoAnalysis;
  edit_plan: ReelEditPlan;
  output: ProcessedReelOutput;
}