# lib/api-zod/src/generated/types/reelEditClip.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReelEditClip {
  segment_id: string;
  source_start_seconds: number;
  source_end_seconds: number;
  order: number;
  selection_reason: string;
  caption_text: string;
  caption_start_seconds: number;
  caption_end_seconds: number;
  reframe_x: number;
  reframe_y: number;
}