# lib/api-zod/src/generated/types/reelEditPlan.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReelEditClip } from './reelEditClip';

export interface ReelEditPlan {
  story_arc: string;
  hook_text: string;
  hook_duration_seconds: number;
  clips: ReelEditClip[];
}