# lib/api-zod/src/generated/types/segmentAnalysis.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SegmentAnalysisLabelsItem } from './segmentAnalysisLabelsItem';

export interface SegmentAnalysis {
  segment_id: string;
  start: number;
  end: number;
  summary: string;
  visual_events: string[];
  labels: SegmentAnalysisLabelsItem[];
  quality_score: number;
  interest_score: number;
  repair_relevance: number;
  transformation_value: number;
}