# lib/api-zod/src/generated/types/videoAnalysis.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SegmentAnalysis } from './segmentAnalysis';
import type { VideoAnalysisArtifacts } from './videoAnalysisArtifacts';

export interface VideoAnalysis {
  video_id: string;
  duration_seconds: number;
  width: number;
  height: number;
  has_audio: boolean;
  scene_change_seconds: number[];
  segments: SegmentAnalysis[];
  artifacts: VideoAnalysisArtifacts;
}