# lib/api-zod/src/generated/types/videoAnalysisArtifacts.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VideoAnalysisArtifacts {
  segments_json: string;
  analysis_json: string;
}