# lib/api-zod/src/generated/types/videoMetadata.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VideoMetadata {
  durationSeconds: number;
  width: number;
  height: number;
  /** @nullable */
  fps: number | null;
  hasAudio: boolean;
}