# lib/api-zod/src/generated/types/videoUploadResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VideoMetadata } from './videoMetadata';

export interface VideoUploadResult {
  videoId: string;
  metadata: VideoMetadata;
}